Move aggregation type to OpenTelemetry.Api
